+++ packages/cms/lib/modules/arguments-form-widgets/index.js
@@ -7,7 +7,7 @@ const eventEmitter  = require('../../../events').emitter;
 
 module.exports = {
   extend: 'openstad-widgets',
-  label: 'Arguments form',
+  label: 'Arguments form old',
   adminOnly: true,
   addFields: [
       /*